feat: integrate API searches
